moved fixes to patches
